test(#3515): split the budget layer into graceful_shutdown_support/budgets.rs (file-size)
The round-6 work pushed `graceful_shutdown_support/mod.rs` to 1584 lines, over the
1500 test threshold, so `file-size` would have FAILed the next lite. Split by
responsibility rather than line count:

* `graceful_shutdown_support/budgets.rs` (931) -- the CLOCK AND THE BUDGETS: stage
  specs, the floor-invariant mapping, the calibration anchors/baselines,
  `Budget`/`calibrated`/`bare`, `StageClock`, the permissive-anchor NOTICE, and
  every unit test that pins their invariants. The tests live with the constants
  they constrain, so a constant cannot be edited without its guard in view.
* `graceful_shutdown_support/mod.rs` (672) -- the CHILD HARNESS: markers, pipes +
  shared transcript, `WaitEnd`, the progress-checked poll, `count_data_db`, the
  bounded read-side SELECT, and the session spawn/ack helpers.
* `graceful_shutdown_tests.rs` (448) -- unchanged: the two subject tests.

Constraints preserved and re-verified: the test TARGET NAME is untouched (still
the only `*.rs` at `cqlite-cli/tests` maxdepth 1 matching graceful, so the
`cli-tests` enumeration and the zero-tests allow-list still resolve), both new
files are inside the existing subdirectory so neither becomes a cargo target, and
the test count is unchanged at 10 (`running 10 tests`). All three files are under
the limit; clippy `-D warnings` clean.
